create card detail konsultasi
